backend/requirements.txt
# Flask and Web Framework
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
werkzeug==2.0.1
requests>=2.27.0

# Twitter API
tweepy==4.12.0

# Data Processing
numpy>=1.22,<1.24
pandas==1.3.3

# NLP and ML
nltk==3.6.3
scikit-learn==0.24.2
transformers==4.26.0
tokenizers==0.12.1
tensorflow==2.12.0
torch==1.13.0
textblob==0.17.1
regex>=2023.0.0

# Environment and Utils
python-dotenv==0.19.0
redis>=4.5.0
markupsafe==2.0.1
jinja2==3.0.1
click==8.0.1

# Database
SQLAlchemy>=1.4.0
sqlalchemy==1.4.47  # Pin to a compatible version

# News API
newsapi-python==0.2.7

# Rate Limiter
flask-limiter==3.3.0
